Const Records: Relax `shape` for Particles
For advanced, highly parallel I/O output, we developed a new
method in ADIOS2 that does not need an initial metadata gather
("JoinedArrays"). To be able to use this mode, we need to relax
the requirements to write a shape for constant records in a species
(particle group), because otherwise we still have to do a collective
gather.

This adds the need for a slight additional read fallback implementation
on the reader side.